# openapi.json
{
  "components": {
    "schemas": {
...
      "CommChannel": {
        "description": "Outbound channels (04 \u00a75.7).",
        "enum": [
          "sms",
          "email"
        ],
        "title": "CommChannel",
        "type": "string"
      },
...
      "ErrorResponse": {
        "description": "OpenAPI shape of the error envelope, so documented error responses match\nwhat the handler emits (keeps the Schemathesis contract test honest).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FarmerEnrollRequest": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "pattern": "^[^\\x00]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "consent_granted": {
            "title": "Consent Granted",
            "type": "boolean"
          },
          "consent_text_version": {
            "pattern": "^[^\\x00]*$",
            "title": "Consent Text Version",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "pattern": "^[^\\x00]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "first_name": {
            "pattern": "^[^\\x00]*$",
            "title": "First Name",
            "type": "string"
          },
          "last_name": {
            "pattern": "^[^\\x00]*$",
            "title": "Last Name",
            "type": "string"
          },
          "number_of_hives": {
            "anyOf": [
              {
                "maximum": 2147483647.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number Of Hives"
          },
          "password": {
            "minLength": 1,
            "pattern": "^[^\\x00]*$",
            "title": "Password",
            "type": "string"
          },
          "phone": {
            "pattern": "^[^\\x00]*$",
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "first_name",
          "last_name",
          "phone",
          "password",
          "consent_granted",
          "consent_text_version"
        ],
        "title": "FarmerEnrollRequest",
        "type": "object"
      },
      "FarmerResponse": {
        "properties": {
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "first_name",
          "last_name",
          "phone",
          "user_id"
        ],
        "title": "FarmerResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RequestValidationErrorResponse": {
        "description": "FastAPI's own 422 body for a request that fails schema validation.\n\nA route that also raises a *domain* 422 (`consent_required`,\n`invalid_code`, ...) declares 422 itself, which replaces FastAPI's default\n422 schema. Without this second shape the declaration would then promise\nonly the envelope, and every malformed body would violate the contract\n(P3b-B: Schemathesis sent a lone NUL byte as a body and found exactly that).",
        "properties": {
          "detail": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "RequestValidationErrorResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerificationConfirmRequest": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/CommChannel"
          },
          "code": {
            "pattern": "^[0-9]{6}$",
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "channel",
          "code"
        ],
        "title": "VerificationConfirmRequest",
        "type": "object"
      },
      "VerificationConfirmedResponse": {
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/CommChannel"
          },
          "verified_at": {
            "format": "date-time",
            "title": "Verified At",
            "type": "string"
          }
        },
        "required": [
          "channel",
          "verified_at"
        ],
        "title": "VerificationConfirmedResponse",
        "type": "object"
      },
      "VerificationSendRequest": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/CommChannel"
          }
        },
        "required": [
          "channel"
        ],
        "title": "VerificationSendRequest",
        "type": "object"
      },
      "VerificationSentResponse": {
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/CommChannel"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "resend_available_at": {
            "format": "date-time",
            "title": "Resend Available At",
            "type": "string"
          }
        },
        "required": [
          "channel",
          "expires_at",
          "resend_available_at"
        ],
        "title": "VerificationSentResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Honey provenance and traceability: hash-chained audit log, Merkle anchoring via OpenTimestamps, PostGIS geospatial records.",
    "title": "ApiChain v2 Backend",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/farmers": {
      "post": {
        "operationId": "enroll_farmer_v2_farmers_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FarmerEnrollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Malformed request body"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Enroll Farmer",
        "tags": [
          "farmers"
        ]
      }
    },
...
    "/v2/farmers/{farmer_id}/verifications": {
      "post": {
        "operationId": "send_verification_code_v2_farmers__farmer_id__verifications_post",
        "parameters": [
          {
            "in": "path",
            "name": "farmer_id",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "title": "Farmer Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationSentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Malformed request body"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/RequestValidationErrorResponse"
                    }
                  ],
                  "title": "Response 422 Send Verification Code V2 Farmers  Farmer Id  Verifications Post"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Verification Code",
        "tags": [
          "verifications"
        ]
      }
    },
    "/v2/farmers/{farmer_id}/verifications/confirm": {
      "post": {
        "operationId": "confirm_verification_code_v2_farmers__farmer_id__verifications_confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "farmer_id",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "title": "Farmer Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationConfirmedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Malformed request body"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/RequestValidationErrorResponse"
                    }
                  ],
                  "title": "Response 422 Confirm Verification Code V2 Farmers  Farmer Id  Verifications Confirm Post"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Confirm Verification Code",
        "tags": [
          "verifications"
        ]
      }
    },
...
  }
}